+++ scripts/devops_tasks/create_coverage.py
@@ -19,8 +19,7 @@ logging.getLogger().setLevel(logging.INFO)
 
 root_dir = os.path.abspath(os.path.join(os.path.abspath(__file__), "..", "..", ".."))
 coverage_dir = os.path.join(root_dir, "_coverage/")
-toxrc = os.path.join(root_dir, "eng","tox", "tox.ini")
-
+coveragerc = os.path.join(root_dir, ".coveragerc")
 
 def collect_tox_coverage_files():
     coverage_version_cmd = [sys.executable, "-m", "coverage", "--version"]
@@ -53,7 +52,7 @@ def collect_tox_coverage_files():
 def generate_coverage_xml():
     if os.path.exists(coverage_dir):
         logging.info("Generating coverage XML")
-        commands = ["coverage", "xml", "-i", "--rcfile", toxrc]
+        commands = ["coverage", "xml", "-i", "--rcfile", coveragerc]
         run_check_call(commands, root_dir, always_exit=False)
     else:
         logging.error("Coverage file is not available in {} to generate coverage XML".format(coverage_dir))