+++ sdk/databox/azure-mgmt-databox/generated_samples/jobs_create_export.py
@@ -75,12 +75,12 @@ def main():
                 },
                 "transferType": "ExportFromAzure",
             },
-            "sku": {"name": "DataBox"},
+            "sku": {"model": "DataBox", "name": "DataBox"},
         },
     ).result()
     print(response)
 
 
-# x-ms-original-file: specification/databox/resource-manager/Microsoft.DataBox/stable/2022-12-01/examples/JobsCreateExport.json
+# x-ms-original-file: specification/databox/resource-manager/Microsoft.DataBox/stable/2025-02-01/examples/JobsCreateExport.json
 if __name__ == "__main__":
     main()