+++ sdk/core/azure-core/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 1.33.0 (2025-03-10)
+## 1.33.0 (Unreleased)
 
 ### Features Added
 
@@ -26,6 +26,10 @@
 - Renamed span attributes in `DistributedTracingPolicy`:
     - "x-ms-client-request-id" is now "az.client_request_id"
     - "x-ms-request-id" is now "az.service_request_id"
+
+### Bugs Fixed
+
+- Fixed an issue where the `traceparent` header was not being set correctly in the `DistributedTracingPolicy`. The `traceparent` header will now set based on the context of the HTTP client span. #40074
 
 ### Other Changes